Moved CellWithChecBox function to DataTableCells file.
